> Sonar > Cambios BBDD

+++ APPWEB-INCOMEMotorCalculoJava-model/src/main/java/com/inditex/rrhh/icmclcwb/model/primary/calcular/entity/Algoritmo.java
@@ -22,7 +22,7 @@ public class Algoritmo {
 
     @Id
     @Column(name = "ID_ALGORITMO")
-    private /* BigInteger */ Long id;
+    private Integer id;
 
     @NotNull
     @Column(name = "ACTIVO", nullable = false)
